Ajuste tabelas detalhes
